Keep only the arrangement the workspace can reach
`panePresentation` took a position and a viewport and resolved them against
each other, but its one caller passed constants: a secondary always present,
a viewport never narrow. It returned its argument. The behaviour it existed
for, a preference the room can override, has nothing to supply it yet, and
its tests asserted that behaviour into a vacuum.

The three-way choice is in `paneState` instead, which is the whole of what
today can produce. The preference and the override come back with the control
that lets someone set a position, and they will arrive with a caller.

`showsSecondary` had no readers at all, which is the thing this branch removed
`pane-surfaces` for. Gone.

`isAppMainView` and `isOverlayView` move across rather than keeping a module
of their own, leaving one place that says what the workspace is showing.
